Remove helios-remote-types dependency on helios-oci
It achieves this by moving ImageUri to helios_util::types which should
be the main place for common types.

It also removes the dependency of legacy and remote crates on mahler by
just re-exporting mahler::Interrupt from the helios-util crate

Change-type: patch
