+++ helios-legacy/Cargo.toml
@@ -16,7 +16,6 @@ version.workspace = true
 helios-util = { path = "../helios-util", version = "0" }
 
 axum.workspace = true
-mahler.workspace = true
 reqwest.workspace = true
 serde.workspace = true
 serde_json.workspace = true
