+++ helios-oci/src/registry.rs
@@ -8,9 +8,9 @@ use serde::{Deserialize, Serialize};
 use thiserror::Error;
 use tracing::debug;
 
-use crate::models::ImageUri;
 use crate::util::http::{InvalidUriError, Uri};
 use crate::util::request::{Get, GetConfig, GetError, RequestConfig};
+use crate::util::types::ImageUri;
 
 // See: https://github.com/balena-io/open-balena-api/blob/master/src/lib/config.ts#L476-L479
 const REGISTRY_TOKEN_EXPIRE_SECONDS: Duration = Duration::from_secs(4 * 3600);
