+++ helios-state/src/models/app.rs
@@ -2,8 +2,8 @@ use mahler::State;
 use serde::{Deserialize, Serialize};
 use std::collections::BTreeMap;
 
+use crate::common_types::Uuid;
 use crate::remote_types::AppTarget as RemoteAppTarget;
-use crate::util::types::Uuid;
 
 use super::release::Release;
