+++ helios-util/Cargo.toml
@@ -17,6 +17,7 @@ axum.workspace = true
 dirs.workspace = true
 fastrand.workspace = true
 mahler.workspace = true
+regex.workspace = true
 reqwest.workspace = true
 serde.workspace = true
 serde_json.workspace = true
