+++ helios-util/src/interrupt.rs
@@ -1,0 +1,2 @@
+// re-export mahler's Interrupt implementation
+pub use mahler::workflow::Interrupt;
