+++ helios-util/src/lib.rs
@@ -2,6 +2,7 @@ pub mod config;
 pub mod crypto;
 pub mod fs;
 pub mod http;
+pub mod interrupt;
 pub mod json;
 pub mod request;
 pub mod types;
