+++ helios-util/src/types/image_uri.rs
@@ -5,6 +5,7 @@ use std::{
     sync::LazyLock,
 };
 
+use mahler::State;
 use regex::Regex;
 use serde::{Deserialize, Serialize};
 use thiserror::Error;
@@ -13,7 +14,7 @@ use thiserror::Error;
 #[error("invalid uri format, expected [domain.tld/]repo/image[:tag][@digest] format, got: {0}")]
 pub struct InvalidImageUriError(String);
 
-#[derive(Debug, Clone, Eq)]
+#[derive(State, Debug, Clone, Eq)]
 pub struct ImageUri {
     registry: Option<String>,
     image: String,
