+++ helios-util/src/types/operating_system.rs
@@ -1,89 +1,6 @@
-use mahler::State;
 use serde::{Deserialize, Serialize};
-use std::ops::Deref;
 use std::{fmt::Display, str::FromStr};
 use thiserror::Error;
-
-use crate::crypto::{ALPHA_NUM, pseudorandom_string};
-
-// Just an alias for more descriptive code
-pub type DeviceType = String;
-
-#[derive(State, Serialize, Deserialize, Debug, Clone, PartialEq, Eq, Hash, PartialOrd, Ord)]
-pub struct Uuid(String);
-
-impl Deref for Uuid {
-    type Target = String;
-
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-impl Default for Uuid {
-    fn default() -> Self {
-        Self(uuid::Uuid::new_v4().simple().to_string())
-    }
-}
-
-impl Display for Uuid {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        self.0.fmt(f)
-    }
-}
-
-impl From<String> for Uuid {
-    fn from(value: String) -> Self {
-        Self(value)
-    }
-}
-
-impl From<&str> for Uuid {
-    fn from(value: &str) -> Self {
-        Self(value.to_owned())
-    }
-}
-
-impl From<Uuid> for String {
-    fn from(value: Uuid) -> Self {
-        value.0
-    }
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Eq, Hash)]
-pub struct ApiKey(String);
-
-impl Deref for ApiKey {
-    type Target = String;
-
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-impl Default for ApiKey {
-    fn default() -> Self {
-        Self(pseudorandom_string(ALPHA_NUM, 32))
-    }
-}
-
-impl Display for ApiKey {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        self.0.fmt(f)
-    }
-}
-
-impl From<String> for ApiKey {
-    fn from(value: String) -> Self {
-        Self(value)
-    }
-}
-
-impl From<ApiKey> for String {
-    fn from(value: ApiKey) -> Self {
-        value.0
-    }
-}
 
 #[derive(Debug, Error)]
 #[error("failed to parse OS version string, expected '<name> [<semver>]': got '{0}'")]
