+++ helios-util/src/types/uuid.rs
@@ -1,0 +1,46 @@
+use serde::{Deserialize, Serialize};
+use std::fmt::Display;
+use std::ops::Deref;
+
+use mahler::State;
+
+#[derive(State, Serialize, Deserialize, Debug, Clone, PartialEq, Eq, Hash, PartialOrd, Ord)]
+pub struct Uuid(String);
+
+impl Deref for Uuid {
+    type Target = String;
+
+    fn deref(&self) -> &Self::Target {
+        &self.0
+    }
+}
+
+impl Default for Uuid {
+    fn default() -> Self {
+        Self(uuid::Uuid::new_v4().simple().to_string())
+    }
+}
+
+impl Display for Uuid {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        self.0.fmt(f)
+    }
+}
+
+impl From<String> for Uuid {
+    fn from(value: String) -> Self {
+        Self(value)
+    }
+}
+
+impl From<&str> for Uuid {
+    fn from(value: &str) -> Self {
+        Self(value.to_owned())
+    }
+}
+
+impl From<Uuid> for String {
+    fn from(value: Uuid) -> Self {
+        value.0
+    }
+}
